libcaption: Fix compiler warnings
CMake build framework 3.0 introduces more strict typing checks, which
require additional code fixes to silence.

* Fixes warnings about implicit casts
* Fixes warnings about missing newlines at the end of files

+++ deps/libcaption/src/cea708.c
@@ -171,19 +171,19 @@ int cea708_render(cea708_t* cea708, uint8_t* data, size_t size)
 {
     int i;
     size_t total = 0;
-    data[0] = cea708->country;
-    data[1] = cea708->provider >> 8;
-    data[2] = cea708->provider >> 0;
+    data[0] = (uint8_t)cea708->country;
+    data[1] = (uint8_t)(cea708->provider >> 8);
+    data[2] = (uint8_t)(cea708->provider >> 0);
     total += 3;
     data += 3;
     size -= 3;
 
     if (t35_provider_atsc == cea708->provider) {
 
-        data[0] = cea708->user_identifier >> 24;
-        data[1] = cea708->user_identifier >> 16;
-        data[2] = cea708->user_identifier >> 8;
-        data[3] = cea708->user_identifier >> 0;
+        data[0] = (uint8_t)(cea708->user_identifier >> 24);
+        data[1] = (uint8_t)(cea708->user_identifier >> 16);
+        data[2] = (uint8_t)(cea708->user_identifier >> 8);
+        data[3] = (uint8_t)(cea708->user_identifier >> 0);
         total += 4;
         data += 4;
         size -= 4;
@@ -214,9 +214,9 @@ int cea708_render(cea708_t* cea708, uint8_t* data, size_t size)
     size -= 2;
 
     for (i = 0; i < (int)cea708->user_data.cc_count; ++i) {
-        data[0] = (cea708->user_data.cc_data[i].marker_bits << 3) | (cea708->user_data.cc_data[i].cc_valid << 2) | cea708->user_data.cc_data[i].cc_type;
+        data[0] = (uint8_t)((cea708->user_data.cc_data[i].marker_bits << 3) | (cea708->user_data.cc_data[i].cc_valid << 2) | cea708->user_data.cc_data[i].cc_type);
         data[1] = cea708->user_data.cc_data[i].cc_data >> 8;
-        data[2] = cea708->user_data.cc_data[i].cc_data >> 0;
+        data[2] = (uint8_t)(cea708->user_data.cc_data[i].cc_data >> 0);
         total += 3;
         data += 3;
         size -= 3;
